# [RFC] Performance improvements of ring buffer processing