--- a/userspace/libscap/scap_bpf.h
+++ b/userspace/libscap/scap_bpf.h
@@ -67,7 +67,9 @@ static inline void scap_bpf_get_buf_pointers(char *buf, uint64_t *phead, uint64_
 	*phead = header->data_head;
 	*ptail = header->data_tail;
 
+	// clang-format off
 	asm volatile("" ::: "memory");
+	// clang-format on
 
 	begin = *ptail % header->data_size;
 	end = *phead % header->data_size;
@@ -154,6 +156,23 @@ static inline int32_t scap_bpf_advance_to_evt(scap_t *handle, uint16_t cpuid, bo
 	return SCAP_SUCCESS;
 }
 
+static inline void scap_bpf_advance_tail(scap_t *handle, uint32_t cpuid)
+{
+	struct perf_event_mmap_page *header;
+	struct scap_device *dev;
+
+	dev = &handle->m_devs[cpuid];
+	header = (struct perf_event_mmap_page *)dev->m_buffer;
+
+	// clang-format off
+	asm volatile("" ::: "memory");
+	// clang-format on
+
+	ASSERT(dev->m_lastreadsize > 0);
+	header->data_tail += dev->m_lastreadsize;
+	dev->m_lastreadsize = 0;
+}
+
 static inline int32_t scap_bpf_readbuf(scap_t *handle, uint32_t cpuid, char **buf, uint32_t *len)
 {
 	struct perf_event_mmap_page *header;
@@ -166,9 +185,7 @@ static inline int32_t scap_bpf_readbuf(scap_t *handle, uint32_t cpuid, char **bu
 	dev = &handle->m_devs[cpuid];
 	header = (struct perf_event_mmap_page *) dev->m_buffer;
 
-	asm volatile("" ::: "memory");
-	header->data_tail += dev->m_lastreadsize;
-
+	ASSERT(dev->m_lastreadsize == 0);
 	scap_bpf_get_buf_pointers((char *) header, &head, &tail, &read_size);
 
 	dev->m_lastreadsize = read_size;